+++ udata_search_service/infrastructure/search_clients.py
@@ -116,6 +116,7 @@ class SearchableReuse(IndexDocument):
     organization = Keyword()
     description = Text(analyzer=dgv_analyzer)
     organization_name = Text(analyzer=dgv_analyzer)
+    organization_badges = Keyword(multi=True)
     owner = Keyword()
 
     class Index:
@@ -148,6 +149,7 @@ class SearchableDataset(IndexDocument):
     description = Text(analyzer=dgv_analyzer)
     organization = Keyword()
     organization_name = Text(analyzer=dgv_analyzer)
+    organization_badges = Keyword(multi=True)
     owner = Keyword()
     schema = Keyword(multi=True)
     topics = Keyword(multi=True)